Merge (#1)
* Added timezone support (#22)

- Added timezone support (using environment variable TZ)
- Added bash to dockerfile

* Configurable device ID (#24)

* Configurable device ID

* Removed dummy device ID

* Updated the API

Updated the API for the new version of the TSSA, and made it work with Python3.

* Device ID fixes

Changed how the device ID is set. Now you can enter a custom device ID when you login, or if it is already set in your settings.py it will use that one.

Removed the need for Google Maps from the API_REQUESTS too.